pyproject.toml
[project]
name = "labq-rag"
version = "0.1.0"
description = "RAG service for LabQ internal knowledge"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # LangChain 1.0+ (2025.10 릴리즈)
    "langchain>=1.0.0",
    "langgraph>=1.0.0",
    "langchain-qdrant>=1.1.0",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=0.3.0",
    # Web Framework
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "python-multipart>=0.0.9",
    # Vector DB & PDF
    "qdrant-client>=1.16.0",
    "pypdf>=6.6.0",
    # Utils
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "langchain-google-genai>=4.2.0",
    "langchain-anthropic>=1.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.14",
]
